Allow models to specifiy conflict keys
fixes #12288

The upsert DAO function generated uses the primary
key as an "ON CONFLICT" match. This might not be the case
if a field uses a unique flag or if that field is also
workgroupable. Postgresql requires an index to be specified.

I tried autodetecting the correct behavior but it failed
on other models. Changing the primary key also did not work.
I added a way to specify the conflict index.
It defaults to the primary key.